requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread==6.0.2
google-auth==2.27.0
httpx==0.26.0
